fix: segemnt scoped artifact uploads
Ensure that the SEGMENT value is not overridden.

If the ENVIRONMENT is not defined, as is common during builds, the
automation `setContent.sh` will force the ENVIRONMENT to be the segment
and the SEGMENT to be `default`.

This fix saves and restores the segment. A future fix will be to
remove the defaulting behaviour as everything should now be defining
ENVIRONMENT and SEGMENT.