feat: CMDB construction
In preparation for converting the library functions to use the
hamlet CLI, refactor properties handling to distinguish it
from CMDB management.

The repo associated with any job is now treated as the source of
pipeline properties files. If the same repo is needed to construct
a CMDB, it is separately loaded.

A number of the previously hardcoded locations are now provided as
default values to function parameters.