fix: env var case sensitivity
Chnage the names and case of the local variables to avoid
issues with Jenkins' handling of case sensitivity in environment
variables.

When setting an environment variable, if an upper case version
aloready exists, Jenkins will set it even if a lower case version
of the variable name is used. This doesn't work well if the
variable is then used in a bash script.