feat: appendItemsToList
Add function to abstract the vagaries of groovy when managing
comma separated lists held in environment variables.

Add the ability for library funnctions to honour environment
overrides in segment properties files. This is initially being
done to permit a segment to use a different hamlet version to others,
but could equally be done to override things like account assignments.